requirements.txt
async-timeout==4.0.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.4
